Fix crash on 64 bit process (#2670)
Fixes #2668

+++ MaterialDesignThemes.Wpf/ScrollViewerAssist.cs
@@ -148,7 +148,7 @@ namespace MaterialDesignThemes.Wpf
                     switch (msg)
                     {
                         case WM_MOUSEHWHEEL:
-                            int tilt = (short)((wParam.ToInt32() >> 16) & 0xFFFF);
+                            int tilt = (short)((wParam.ToInt64() >> 16) & 0xFFFF);
                             OnMouseTilt(tilt);
                             return (IntPtr)1;
                     }
